CSCETSIN-377 python REST API functionality for qvain light moved to their own files
